Add a CheckLogin higher order component. 😎
This is an example of the Higher Order Component patterne from the react docs. It is okay if it
freaks out out at first! The idea is that we have some component CheckLogin that just takes
another component as a prop and decides whether or not to render it. If the user is logged in,
it will render it. If the user isn't logged in, it will redirect them to the login page.
Tada! 🎉

Now we use it in the DogContainer to make sure that you can't get to the new dog form unless
you are already logged in. Higher order components rule and we'll see more of them soon...